+++ icons/src/main/java/org/patternfly/icon/PredefinedIcon.java
@@ -49,6 +49,40 @@ public final class PredefinedIcon implements
         SVGElementDataMethods<SVGElement, PredefinedIcon>,
         SVGElementStyleMethods<SVGElement, PredefinedIcon> {
 
+    // ------------------------------------------------------ factory
+
+    /**
+     * Constructs a predefined SVG icon based on the provided icon name. The method supports multiple icon groups, such as
+     * "fab", "far", "fas", and "patternfly". If the name contains a group prefix (e.g., "fab.iconName"), the method will use
+     * that group; otherwise, it defaults to the "fas" group.
+     * <p>
+     * Normally you should use the static methods from {@link IconSets} to get predefined icons. Only use this method if you
+     * need to get an icon based on the icon name provided as string.
+     *
+     * @param name the name of the icon, optionally prefixed with the group name (e.g., "fab.iconName"). If no group is
+     *             specified, "fas" is used by default.
+     * @return a {@code PredefinedIcon} instance representing the specified icon.
+     * @throws IllegalArgumentException if the group or icon name is unknown.
+     */
+    public static PredefinedIcon predefinedIcon(String name) {
+        String group = "fas";
+        String iconName = name;
+        if (name.contains(".")) {
+            group = name.substring(0, name.indexOf('.'));
+            iconName = name.substring(name.indexOf('.') + 1);
+        }
+        IconSpec iconSpec = switch (group) {
+            case "fab" -> IconSpecs.fab.valueOf(iconName).iconSpec;
+            case "far" -> IconSpecs.far.valueOf(iconName).iconSpec;
+            case "fas" -> IconSpecs.fas.valueOf(iconName).iconSpec;
+            case "patternfly" -> IconSpecs.patternfly.valueOf(iconName).iconSpec;
+            default -> throw new IllegalArgumentException("Unknown icon group/name: '" + name + "'");
+        };
+        return new PredefinedIcon(iconSpec);
+    }
+
+    // ------------------------------------------------------ instance
+
     public final IconSpec iconSpec;
     private final SVGElement element;