+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+- Add `PredefinedIcon.predefinedIcon(String iconName)`
+
 ## [0.4.2] - 2025-12-17
 
 ### Added